requirements.txt
flask
flask-socketio
gevent
gevent-websocket
requests
python-dotenv
gunicorn
urllib3
